requirements.txt
boto3==1.34.131
requests==2.28.1
urllib3==1.26.6
python_dotenv
mistralai
pypandoc
bs4
pydantic
markdown
fastapi
nest-asyncio
uvicorn